implement CRUD operations for Review Entity
Bug fixes

+++ src/main/java/com/jemo/RestaurantReviewPortal/menu/MenuService.java
@@ -2,6 +2,7 @@ package com.jemo.RestaurantReviewPortal.menu;
 
 import com.jemo.RestaurantReviewPortal.restaurant.Restaurant;
 import com.jemo.RestaurantReviewPortal.restaurant.RestaurantRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -25,21 +26,27 @@ public class MenuService {
         return menuRepository.findAllByRestaurantId(restaurantId);
     }
 
+    @Transactional
     public Boolean deleteById(Long restaurantId, Long menuId) {
         Menu menu = findMenuByRestaurantIdAndMenuId(restaurantId, menuId);
-        menuRepository.delete(menu);
+        if(menu != null) {
+            menuRepository.delete(menu);
+        }
+
+        // confirm deletion
         if(findMenuByRestaurantIdAndMenuId(restaurantId, menuId) == null) {
             return true;
         }
         return false;
     }
 
+    @Transactional
     public Boolean updateById(Long restaurantId, Long menuId, MenuRequest menuRequest) {
         Menu menu = findMenuByRestaurantIdAndMenuId(restaurantId, menuId);
         if(menu.getId() != null) {
             Menu updatedMenu = Menu.builder()
                     .id(menu.getId())
-                    .restaurantId(menu.getRestaurantId())
+                    .restaurant(menu.getRestaurant())
                     .description(menuRequest.description() != null ? menuRequest.description() : menu.getDescription())
                     .name(menuRequest.name() != null ? menuRequest.name() : menu.getName())
                     .build();
@@ -49,6 +56,7 @@ public class MenuService {
         return false;
     }
 
+    @Transactional
     public Boolean createMenu(Long restaurantId, MenuRequest menuRequest) {
         // check if menu exists
         Menu menuExists = menuRepository.findByRestaurantIdAndName(restaurantId, menuRequest.name());
@@ -62,7 +70,7 @@ public class MenuService {
         }
         // create menu
         Menu newMenu = Menu.builder()
-                 .restaurantId(restaurantId)
+                 .restaurant(restaurant)
                  .name(menuRequest.name())
                  .description(menuRequest.description())
                  .build();
